@@ -1,4 +1,5 @@
-﻿using System;
+﻿using ImageMagick;
+using System;
 using System.ComponentModel;
 using System.Drawing;
 using UndertaleModLib.Util;
@@ -146,24 +147,21 @@ public class UndertaleTexturePageItem : UndertaleNamedResource, INotifyPropertyC
     /// </summary>
     /// <param name="replaceImage">The new image that shall be applied to this texture page item.</param>
     /// <param name="disposeImage">Whether to dispose <paramref name="replaceImage"/> afterwards.</param>
-    public void ReplaceTexture(Image replaceImage, bool disposeImage = true)
+    public void ReplaceTexture(MagickImage replaceImage, bool disposeImage = true)
     {
-        Image finalImage = TextureWorker.ResizeImage(replaceImage, SourceWidth, SourceHeight);
+        // Resize image to bounds on texture page
+        IMagickImage<byte> finalImage = TextureWorker.ResizeImage(replaceImage, SourceWidth, SourceHeight);
 
         // Apply the image to the TexturePage.
         lock (TexturePage.TextureData)
         {
-            TextureWorker worker = new TextureWorker();
-            Bitmap embImage = worker.GetEmbeddedTexture(TexturePage); // Use SetPixel if needed.
+            using TextureWorker worker = new();
+            MagickImage embImage = worker.GetEmbeddedTexture(TexturePage);
 
-            Graphics g = Graphics.FromImage(embImage);
-            g.CompositingMode = System.Drawing.Drawing2D.CompositingMode.SourceCopy;
-            g.DrawImage(finalImage, SourceX, SourceY);
-            g.Dispose();
+            embImage.Composite(finalImage, SourceX, SourceY, CompositeOperator.Copy);
 
-            TexturePage.TextureData.TextureBlob = TextureWorker.GetImageBytes(embImage);
-
-            worker.Cleanup();
+            // Replace original texture with the new version, in the original texture format
+            TexturePage.TextureData.Image = GMImage.FromMagickImage(embImage).ConvertToFormat(TexturePage.TextureData.Image.Format);
         }
 
         TargetWidth = (ushort)replaceImage.Width;
