@@ -1,59 +1,54 @@
-﻿using System;
+﻿using ImageMagick;
+using System;
 using System.Collections.Generic;
 using System.Drawing;
-using System.Drawing.Drawing2D;
 using System.Drawing.Imaging;
 using System.IO;
 using UndertaleModLib.Models;
 
 namespace UndertaleModLib.Util
 {
-    public class TextureWorker
+    /// <summary>
+    /// Helper class used to manage and cache textures.
+    /// </summary>
+    public class TextureWorker : IDisposable
     {
-        private Dictionary<UndertaleEmbeddedTexture, Bitmap> embeddedDictionary = new Dictionary<UndertaleEmbeddedTexture, Bitmap>();
-        private static readonly ImageConverter _imageConverter = new ImageConverter();
+        private Dictionary<UndertaleEmbeddedTexture, MagickImage> embeddedDictionary = [];
 
-        // Cleans up all the images when usage of this worker is finished.
-        // Should be called when a TextureWorker will never be used again.
-        public void Cleanup()
-        {
-            foreach (Bitmap img in embeddedDictionary.Values)
-                img.Dispose();
-            embeddedDictionary.Clear();
-        }
-
-        public Bitmap GetEmbeddedTexture(UndertaleEmbeddedTexture embeddedTexture)
+        public MagickImage GetEmbeddedTexture(UndertaleEmbeddedTexture embeddedTexture)
         {
             lock (embeddedDictionary)
             {
                 if (!embeddedDictionary.ContainsKey(embeddedTexture))
-                    embeddedDictionary[embeddedTexture] = GetImageFromByteArray(embeddedTexture.TextureData.TextureBlob);
+                {
+                    embeddedDictionary[embeddedTexture] = embeddedTexture.TextureData.Image.GetMagickImage();
+                }
                 return embeddedDictionary[embeddedTexture];
             }
         }
 
-        public void ExportAsPNG(UndertaleTexturePageItem texPageItem, string FullPath, string imageName = null, bool includePadding = false)
+        public void ExportAsPNG(UndertaleTexturePageItem texPageItem, string fullPath, string imageName = null, bool includePadding = false)
         {
-            SaveImageToFile(FullPath, GetTextureFor(texPageItem, imageName != null ? imageName : Path.GetFileNameWithoutExtension(FullPath), includePadding));
+            SaveImageToFile(fullPath, GetTextureFor(texPageItem, imageName ?? Path.GetFileNameWithoutExtension(fullPath), includePadding));
         }
 
-        public Bitmap GetTextureFor(UndertaleTexturePageItem texPageItem, string imageName, bool includePadding = false)
+        public IMagickImage<byte> GetTextureFor(UndertaleTexturePageItem texPageItem, string imageName, bool includePadding = false)
         {
             int exportWidth = texPageItem.BoundingWidth; // sprite.Width
             int exportHeight = texPageItem.BoundingHeight; // sprite.Height
-            Bitmap embeddedImage = GetEmbeddedTexture(texPageItem.TexturePage);
+            MagickImage embeddedImage = GetEmbeddedTexture(texPageItem.TexturePage);
 
             // Sanity checks.
             if (includePadding && ((texPageItem.TargetWidth > exportWidth) || (texPageItem.TargetHeight > exportHeight)))
                 throw new InvalidDataException(imageName + "'s texture is larger than its bounding box!");
 
             // Create a bitmap representing that part of the texture page.
-            Bitmap resultImage = null;
+            IMagickImage<byte> resultImage = null;
             lock (embeddedImage)
             {
                 try
                 {
-                    resultImage = embeddedImage.Clone(new Rectangle(texPageItem.SourceX, texPageItem.SourceY, texPageItem.SourceWidth, texPageItem.SourceHeight), PixelFormat.DontCare);
+                    resultImage = embeddedImage.Clone(texPageItem.SourceX, texPageItem.SourceY, texPageItem.SourceWidth, texPageItem.SourceHeight);
                 }
                 catch (OutOfMemoryException)
                 {
@@ -63,88 +58,76 @@ namespace UndertaleModLib.Util
 
             // Resize the image, if necessary.
             if ((texPageItem.SourceWidth != texPageItem.TargetWidth) || (texPageItem.SourceHeight != texPageItem.TargetHeight))
+            {
+                IMagickImage<byte> original = resultImage;
                 resultImage = ResizeImage(resultImage, texPageItem.TargetWidth, texPageItem.TargetHeight);
+                original.Dispose();
+            }
 
             // Put it in the final holder image.
-            Bitmap returnImage = resultImage;
+            IMagickImage<byte> returnImage = resultImage;
             if (includePadding)
             {
-                returnImage = new Bitmap(exportWidth, exportHeight);
-                Graphics g = Graphics.FromImage(returnImage);
-                g.DrawImage(resultImage, new Rectangle(texPageItem.TargetX, texPageItem.TargetY, resultImage.Width, resultImage.Height), new Rectangle(0, 0, resultImage.Width, resultImage.Height), GraphicsUnit.Pixel);
-                g.Dispose();
+                returnImage = new MagickImage(MagickColor.FromRgba(0, 0, 0, 0), exportWidth, exportHeight);
+                returnImage.Composite(resultImage, texPageItem.TargetX, texPageItem.TargetY, CompositeOperator.Copy);
             }
 
             return returnImage;
         }
 
-        public static Bitmap ReadImageFromFile(string filePath)
+        public static MagickImage ReadBGRAImageFromFile(string filePath)
         {
-            return GetImageFromByteArray(File.ReadAllBytes(filePath));
-        }
-
-        // Grabbed from https://stackoverflow.com/questions/3801275/how-to-convert-image-to-byte-array/16576471#16576471
-        public static Bitmap GetImageFromByteArray(byte[] byteArray)
-        {
-            Bitmap bm = (Bitmap)_imageConverter.ConvertFrom(byteArray);
-
-            if (bm != null && (bm.HorizontalResolution != (int)bm.HorizontalResolution ||
-                               bm.VerticalResolution != (int)bm.VerticalResolution))
+            MagickReadSettings settings = new()
             {
-                // Correct a strange glitch that has been observed in the test program when converting 
-                //  from a PNG file image created by CopyImageToByteArray() - the dpi value "drifts" 
-                //  slightly away from the nominal integer value
-                bm.SetResolution((int)(bm.HorizontalResolution + 0.5f),
-                                 (int)(bm.VerticalResolution + 0.5f));
-            }
-
-            return bm;
+                ColorSpace = ColorSpace.sRGB,
+            };
+            MagickImage image = new(filePath, settings);
+            image.Alpha(AlphaOption.Set);
+            image.Format = MagickFormat.Bgra;
+            image.SetCompression(CompressionMethod.NoCompression);
+            return image;
         }
 
         // This should perform a high quality resize.
-        // Grabbed from https://stackoverflow.com/questions/1922040/how-to-resize-an-image-c-sharp
-        public static Bitmap ResizeImage(Image image, int width, int height)
+        public static IMagickImage<byte> ResizeImage(IMagickImage<byte> image, int width, int height)
         {
-            var destRect = new Rectangle(0, 0, width, height);
-            var destImage = new Bitmap(width, height);
-
-            destImage.SetResolution(image.HorizontalResolution, image.VerticalResolution);
-
-            using (var graphics = Graphics.FromImage(destImage))
+            if (image.Width == width && image.Height == height)
             {
-                graphics.CompositingMode = CompositingMode.SourceCopy;
-                graphics.CompositingQuality = CompositingQuality.HighQuality;
-                graphics.InterpolationMode = InterpolationMode.HighQualityBicubic;
-                graphics.SmoothingMode = SmoothingMode.HighQuality;
-                graphics.PixelOffsetMode = PixelOffsetMode.HighQuality;
-
-                using (var wrapMode = new ImageAttributes())
-                {
-                    wrapMode.SetWrapMode(WrapMode.TileFlipXY);
-                    graphics.DrawImage(image, destRect, 0, 0, image.Width, image.Height, GraphicsUnit.Pixel, wrapMode);
-                }
+                return image.Clone();
             }
 
-            return destImage;
+            IMagickImage<byte> newImage = image.Clone();
+            newImage.InterpolativeResize(width, height, PixelInterpolateMethod.Bilinear);
+            return newImage;
         }
 
-        public static byte[] ReadMaskData(string filePath)
+        public static byte[] ReadMaskData(string filePath, int requiredWidth, int requiredHeight)
         {
-            Bitmap image = ReadImageFromFile(filePath);
-            List<byte> bytes = new List<byte>();
+            MagickImage image = ReadBGRAImageFromFile(filePath);
+            if (image.Width != requiredWidth || image.Height != requiredHeight)
+            {
+                throw new Exception($"{filePath} is not the proper size to be imported! The proper dimensions are width: {requiredWidth} px, height: {requiredHeight} px.");
+            }
 
-            int enableColor = Color.White.ToArgb();
+            IPixelCollection<byte> pixels = image.GetPixels();
+            List<byte> bytes = [];
+
+            IMagickColor<byte> enableColor = MagickColor.FromRgba(255, 255, 255, 255);
             for (int y = 0; y < image.Height; y++)
             {
                 for (int xByte = 0; xByte < (image.Width + 7) / 8; xByte++)
                 {
                     byte fullByte = 0x00;
                     int pxStart = (xByte * 8);
-                    int pxEnd = Math.Min(pxStart + 8, (int) image.Width);
+                    int pxEnd = Math.Min(pxStart + 8, image.Width);
 
                     for (int x = pxStart; x < pxEnd; x++)
-                        if (image.GetPixel(x, y).ToArgb() == enableColor) // Don't use Color == OtherColor, it doesn't seem to give us the type of equals comparison we want here.
+                    {
+                        if (pixels.GetPixel(x, y).ToColor().Equals(enableColor))
+                        {
                             fullByte |= (byte)(0b1 << (7 - (x - pxStart)));
+                        }
+                    }
 
                     bytes.Add(fullByte);
                 }
@@ -154,25 +137,13 @@ namespace UndertaleModLib.Util
             return bytes.ToArray();
         }
 
-        public static byte[] ReadTextureBlob(string filePath)
-        {
-            Image.FromFile(filePath).Dispose(); // Make sure the file is valid image.
-            return File.ReadAllBytes(filePath);
-        }
-
-        public static void SaveEmptyPNG(string FullPath, int width, int height)
-        {
-            var blackImage = new Bitmap(width, height);
-            for (int x = 0; x < width; x++)
-                for (int y = 0; y < height; y++)
-                    blackImage.SetPixel(x, y, Color.Black);
-            SaveImageToFile(FullPath, blackImage);
-        }
-
-        public static Bitmap GetCollisionMaskImage(UndertaleSprite sprite, UndertaleSprite.MaskEntry mask)
+        public static IMagickImage<byte> GetCollisionMaskImage(UndertaleSprite sprite, UndertaleSprite.MaskEntry mask)
         {
             byte[] maskData = mask.Data;
-            Bitmap bitmap = new Bitmap((int)sprite.Width, (int)sprite.Height, PixelFormat.Format32bppArgb); // Ugh. I want to use 1bpp, but for some BS reason C# doesn't allow SetPixel in that mode.
+            MagickImage bitmap = new MagickImage(MagickColor.FromRgba(0, 0, 0, 255), (int)sprite.Width, (int)sprite.Height);
+            IPixelCollection<byte> pixels = bitmap.GetPixels();
+            ReadOnlySpan<byte> black = MagickColor.FromRgba(0, 0, 0, 255).ToByteArray().AsSpan();
+            ReadOnlySpan<byte> white = MagickColor.FromRgba(255, 255, 255, 255).ToByteArray().AsSpan();
 
             for (int y = 0; y < sprite.Height; y++)
             {
@@ -181,7 +152,7 @@ namespace UndertaleModLib.Util
                 {
                     byte temp = maskData[rowStart + (x / 8)];
                     bool pixelBit = (temp & (0b1 << (7 - (x % 8)))) != 0b0;
-                    bitmap.SetPixel(x, y, pixelBit ? Color.White : Color.Black);
+                    pixels.SetPixel(x, y, pixelBit ? white : black);
                 }
             }
 
@@ -190,28 +161,38 @@ namespace UndertaleModLib.Util
 
         public static void ExportCollisionMaskPNG(UndertaleSprite sprite, UndertaleSprite.MaskEntry mask, string fullPath)
         {
-            SaveImageToFile(fullPath, GetCollisionMaskImage(sprite, mask));
+            using var image = GetCollisionMaskImage(sprite, mask);
+            SaveImageToFile(fullPath, image);
         }
 
-        public static byte[] GetImageBytes(Image image, bool disposeImage = true)
+        public static void SaveImageToFile(string fullPath, IMagickImage<byte> image, bool disposeImage = true)
         {
-            using (var ms = new MemoryStream())
+            using (var stream = new FileStream(fullPath, FileMode.Create))
             {
-                image.Save(ms, image.RawFormat);
-                byte[] result = ms.ToArray();
-                if (disposeImage)
-                    image.Dispose();
-                return result;
+                image.Write(stream, MagickFormat.Png32);
+            }
+            if (disposeImage)
+            {
+                image.Dispose();
             }
         }
 
-        public static void SaveImageToFile(string FullPath, Image image, Boolean disposeImage = true)
+        /// <summary>
+        /// This should be called when the TextureWorker is no longer going to be used; this frees image data.
+        /// </summary>
+        public void Dispose()
         {
-            var stream = new FileStream(FullPath, FileMode.Create);
-            image.Save(stream, ImageFormat.Png);
-            stream.Close();
-            if (disposeImage)
-                image.Dispose();
+            if (embeddedDictionary is not null)
+            {
+                foreach (MagickImage img in embeddedDictionary.Values)
+                {
+                    img.Dispose();
+                }
+                embeddedDictionary.Clear();
+                embeddedDictionary = null;
+            }
+
+            GC.SuppressFinalize(this);
         }
     }
 }
